third_party/blink/web_tests/html/dialog/closewatcher-skip-cancel-usecounter.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="https://bugs.chromium.org/p/chromium/issues/detail?id=1383001">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<dialog id=dialog></dialog>

<script>
test(() => {
  dialog.showModal();
  eventSender.keyDown('Escape');
  assert_false(internals.isUseCounted(document, 4422));
  assert_false(internals.isUseCounted(document, 4423));

  dialog.showModal();
  dialog.addEventListener('cancel', () => console.log('hello world'));
  eventSender.keyDown('Escape');
  assert_true(internals.isUseCounted(document, 4422));
  assert_false(internals.isUseCounted(document, 4423));

  dialog.showModal();
  dialog.addEventListener('cancel', event => event.preventDefault());
  eventSender.keyDown('Escape');
  assert_true(internals.isUseCounted(document, 4422));
  assert_true(internals.isUseCounted(document, 4423));
});
</script>
